rabbitmq-c: add version 0.12.0, update openssl
