+++ recipes/rabbitmq-c/all/conandata.yml
@@ -1,4 +1,7 @@
 sources:
+  "0.12.0":
+    url: "https://github.com/alanxz/rabbitmq-c/archive/v0.12.0.tar.gz"
+    sha256: "ded0f5088450bc52f288e23f01e5dfc52ea812c020d81fd124d74736da9427b3"
   "0.11.0":
     url: "https://github.com/alanxz/rabbitmq-c/archive/v0.11.0.tar.gz"
     sha256: "437d45e0e35c18cf3e59bcfe5dfe37566547eb121e69fca64b98f5d2c1c2d424"
