+++ recipes/rabbitmq-c/all/conanfile.py
@@ -3,7 +3,7 @@ from conan.tools.cmake import CMake, CMakeDeps, CMakeToolchain, cmake_layout
 from conan.tools.files import copy, get, rmdir
 import os
 
-required_conan_version = ">=1.46.0"
+required_conan_version = ">=1.53.0"
 
 
 class RabbitmqcConan(ConanFile):
@@ -12,7 +12,7 @@ class RabbitmqcConan(ConanFile):
     url = "https://github.com/conan-io/conan-center-index"
     homepage = "https://github.com/alanxz/rabbitmq-c"
     description = "This is a C-language AMQP client library for use with v2.0+ of the RabbitMQ broker."
-    topics = ("rabbitmq-c", "rabbitmq", "message queue")
+    topics = ("rabbitmq", "message queue")
 
     settings = "os", "arch", "compiler", "build_type"
     options = {
@@ -32,19 +32,13 @@ class RabbitmqcConan(ConanFile):
 
     def configure(self):
         if self.options.shared:
-            del self.options.fPIC
-        try:
-            del self.settings.compiler.libcxx
-        except Exception:
-            pass
-        try:
-            del self.settings.compiler.cppstd
-        except Exception:
-            pass
+            self.options.rm_safe("fPIC")
+        self.settings.rm_safe("compiler.libcxx")
+        self.settings.rm_safe("compiler.cppstd")
 
     def requirements(self):
         if self.options.ssl:
-            self.requires("openssl/1.1.1q")
+            self.requires("openssl/1.1.1s")
 
     def layout(self):
         cmake_layout(self, src_folder="src")
