+++ recipes/rabbitmq-c/config.yml
@@ -1,8 +1,9 @@
 versions:
+  "0.12.0":
+    folder: all
   "0.11.0":
     folder: all
   "0.10.0":
     folder: all
   "0.9.0":
     folder: all
-
